requirements.txt
streamlit>=1.40.0
openai>=1.0
google-generativeai>=0.8.0
pandas>=2.0
openpyxl>=3.1
lxml>=5.0
plotly>=5.18
python-dotenv>=1.0
Pillow>=10.0
streamlit-ace>=0.1.1
